Fallback to hex on invalid label names
Sometimes human readable label names are longer than allowed in
kubernetes. This commit will fall back to the hex encoded vendor and
device IDs.
Additionally it will fall back if the device name could not be found in
the `usbid.Vendor` map.
Lastly some small unit test for the label key creation is added and some
refactoring to make that possible.